Move `index.js` to `modules/Instabug.js`
