+++ package.json
@@ -7,7 +7,7 @@
   "homepage": "https://www.instabug.com/platforms/react-native",
   "bugs": "https://github.com/Instabug/Instabug-React-Native/issues",
   "license": "MIT",
-  "main": "src/index.js",
+  "main": "src/modules/Instabug.js",
   "types": "src/index.d.ts",
   "keywords": [
     "react-native",
